refactor(llama): split model out
